+++ src/apiClient/types/api-location-generate-post-request.ts
@@ -93,6 +93,12 @@ export interface ApiLocationGeneratePostRequest {
      */
     'createLongDescription'?: boolean;
     /**
+     * The number of paragraphs to produce in the output when using a long description
+     * @type {number}
+     * @memberof ApiLocationGeneratePostRequest
+     */
+    'longDescriptionParagraphs'?: number;
+    /**
      * The styles of names to use
      * @type {Array<string>}
      * @memberof ApiLocationGeneratePostRequest